requirements.txt
Flask
opencv-python
Werkzeug
darknet-python
torch
matplotlib